pyproject.toml
[project]
name = "cfevals"
version = "0.1.0"
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "chronos-forecasting",
    "datasets",
    "matplotlib",
    "neuralforecast",
    "numpy",
    "openai>=1.0.0",
    "pandas",
    "pandas_datareader",
    "pytest",
    "pyyaml",
    "statsforecast",
    "timesfm",
]

[project.urls]
repository = "https://github.com/openai/evals"

[project.optional-dependencies]
formatters = ["black", "isort", "autoflake", "ruff"]

torch = ["torch"]

[project.scripts]
cfeval = "cfevals.cli.cfeval:main"
cfevalset = "cfevals.cli.cfevalset:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cfevals", "cfevals.*"]

[tool.setuptools.package-data]
cfevals = ["registry/**/*.yaml"]

[tool.ruff]
# Allow lines to be as long as 767 characters, this is currently the longest line
# TODO: This should be brought down in the future
line-length = 767
